Annoyances: silence intercom.com push notification
Sites like www.alphaflow.com (as of 2017-11-14T20:52Z, see in an incognito window) and any others using Intercom's Engage product will pop up a loud autoplaying audio notification on the load of their widget. The widget and the message themselves are useful (as this is usually used for B2B product support) but the audio is incredibly disruptive (especially in work environments).