Improved matching of related domains
On transip.* domains, e.g. transip.be or transip.eu, the following filter in cookies_general.txt prevents the site from accepting the user's acknowledgement of the cookie policy:
/cookie-consent/*^$domain=~argeweb.nl

There is an exception filter in cookies_whitelist.txt for just one of these domains:
@@||transip.nl/services/cookie-consent/

With the large number of related domains, I suggest a wildcard filter.