Suppress Google FLoC on all websites when Tracking Protection filter is enabled
Google started experimenting with FLoC in Chrome browser. This new API is supposed to replace third-party cookies and allow learning about users behavior in a more privacy preserving way than it was.

We are pretty sure that people who use AdGuard Tracking Protection filter are not okay with websites learning about their behavioral groups regardless of how preserving it is.

There's an article by EFF explaining issues of this new API and we agree with every point there:
https://www.eff.org/deeplinks/2021/03/googles-floc-terrible-idea