youtube.com - Social
I modified a filter to remove every social icon except for the "Embed" icon in the share panel. The share panel in the shorts player doesn't have an embed button and a social icon (WhatsApp if you're logged out or Facebook if you're logged in) wasn't being affected by the old filter. With this fix it should work both on the normal and the shorts player.